requirements.txt
# Bono AI Fashion Catalog Pipeline

google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
httpx>=0.25.0
aiohttp>=3.9.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
watchdog>=3.0.0
aiofiles>=23.0.0
tenacity>=8.2.0
structlog>=23.2.0
click>=8.1.0
tqdm>=4.66.0
